Extrae modelos de comida y menu
