+++ src/app/menu/menu.model.ts
@@ -1,0 +1,9 @@
+import {Comida} from '../comida/comida.model';
+
+export type Menu = {
+  id: number
+  nombre: string;
+  precio: number;
+  vegetariano: boolean;
+  comidas: Comida[];
+}
